test: address process isolation review findings
